Rename AuthenticateSubject to ValidateCredential
AuthenticateSubject is sort of a clunky name for what this RPC
actually does, which is validating credentials. This commit renames it
to ValidateCredential and updates the response with subject_id as a
field returned directly instead of arbitrary claims (inspired by
SPIFFE, which has similar APIs).

This is a breaking change.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>